requirements.txt
imageio>=2.34,<3.0
imageio-ffmpeg>=0.5,<1.0
librosa>=0.10,<1.0
numpy>=1.26,<3.0
Pillow>=10.0,<12.0
pydub>=0.25,<1.0
